Update fmod to use std::fmod instead of std::fmodf
Replaced the single-precision std::fmodf with the double-precision std::fmod in the fmod function implementation. This may affect floating-point precision and type compatibility for modulo operations.

+++ sdk/geUtilities/include/geMath.h
@@ -162,7 +162,7 @@ namespace geEngineSDK {
 
     GE_NODISCARD static FORCEINLINE float
     fmod(float X, float Y) {
-      return std::fmodf(X, Y);
+      return std::fmod(X, Y);
     }
 
     GE_NODISCARD static FORCEINLINE float
